Extract `rights` into a model used by `User`
