+++ src/models/mod.rs
@@ -6,7 +6,8 @@ pub use download::VersionDownload;
 pub use self::follow::Follow;
 pub use self::keyword::{CrateKeyword, Keyword};
 pub use self::krate::{Crate, CrateDownload, NewCrate};
-pub use owner::{CrateOwner, NewTeam, Owner, OwnerKind, Rights, Team};
+pub use owner::{CrateOwner, NewTeam, Owner, OwnerKind, Team};
+pub use self::rights::Rights;
 pub use user::{Email, NewUser, User};
 pub use token::ApiToken;
 pub use version::{NewVersion, Version};
@@ -18,3 +19,4 @@ mod category;
 mod follow;
 mod keyword;
 pub mod krate;
+mod rights;
